Move StatsPluginInterface to plugins package.
