+++ libs/stats/odc/stats/plugins/__init__.py
@@ -1,1 +1,6 @@
-from ._base import resolve, import_all, register
+from ._base import (
+    resolve,
+    import_all,
+    register,
+    StatsPluginInterface
+)
