+++ libs/stats/odc/stats/plugins/_base.py
@@ -1,8 +1,37 @@
 import pydoc
-from typing import Callable, Dict
+from abc import ABC, abstractmethod
+from typing import Callable, Dict, Tuple, Sequence, Optional
 from functools import partial
 
-from odc.stats.model import StatsPluginInterface
+import xarray as xr
+from datacube.model import Dataset
+from datacube.utils.geometry import GeoBox
+
+class StatsPluginInterface(ABC):
+    NAME = "*unset*"
+    SHORT_NAME = ""
+    VERSION = "0.0.0"
+    PRODUCT_FAMILY = "statistics"
+
+    @property
+    @abstractmethod
+    def measurements(self) -> Tuple[str, ...]:
+        pass
+
+    @abstractmethod
+    def input_data(self, datasets: Sequence[Dataset], geobox: GeoBox) -> xr.Dataset:
+        pass
+
+    @abstractmethod
+    def reduce(self, xx: xr.Dataset) -> xr.Dataset:
+        pass
+
+    def rgba(self, xx: xr.Dataset) -> Optional[xr.DataArray]:
+        """
+        Given result of ``.reduce(..)`` optionally produce RGBA preview image
+        """
+        return None
+
 
 PluginFactory = Callable[..., StatsPluginInterface]
 
@@ -50,3 +79,4 @@ def import_all():
             importlib.import_module(mod)
         except ModuleNotFoundError:
             pass
+
